# docs/openapi.v1.json
{
  "components": {
    "schemas": {
      "ActorOut": {
        "properties": {
          "responsibility": {
            "title": "Responsibility",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role",
          "responsibility"
        ],
        "title": "ActorOut",
        "type": "object"
      },
...
      "DecisionRequest": {
        "additionalProperties": false,
        "properties": {
          "actor": {
            "description": "The named human who decided.",
            "minLength": 2,
            "title": "Actor",
            "type": "string"
          },
          "decision": {
            "$ref": "#/components/schemas/DecisionValue"
          },
          "reason": {
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          },
          "role": {
            "description": "Must appear in the stage authority matrix.",
            "minLength": 2,
            "title": "Role",
            "type": "string"
          },
          "signature": {
            "minLength": 8,
            "title": "Signature",
            "type": "string"
          }
        },
        "required": [
          "actor",
          "role",
          "decision",
          "reason",
          "signature"
        ],
        "title": "DecisionRequest",
        "type": "object"
      },
      "DecisionResult": {
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "boolean"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "guardrail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guardrail"
          },
          "principal": {
            "description": "The authenticated identity written into the ledger.",
            "title": "Principal",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "accepted",
          "status",
          "detail",
          "principal"
        ],
        "title": "DecisionResult",
        "type": "object"
      },
      "DecisionValue": {
        "enum": [
          "approve",
          "reject"
        ],
        "title": "DecisionValue",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_RunSummary_": {
        "properties": {
          "has_more": {
            "default": false,
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Pass back as `cursor` for the next page. Null on the last page.",
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[RunSummary]",
        "type": "object"
      },
...
      "RunCreate": {
        "additionalProperties": false,
        "properties": {
          "discovery_session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Alternative to `workflow_id`: start the workflow this session resolved.",
            "title": "Discovery Session Id"
          },
          "evidence": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Attached to the first stage as recorded evidence for the run.",
            "title": "Evidence"
          },
          "region": {
            "anyOf": [
              {
                "maxLength": 30,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "title": {
            "maxLength": 250,
            "minLength": 3,
            "title": "Title",
            "type": "string"
          },
          "workflow_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "A compiled workflow to start.",
            "title": "Workflow Id"
          }
        },
        "required": [
          "title"
        ],
        "title": "RunCreate",
        "type": "object"
      },
      "RunDetail": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "organization_id": {
            "title": "Organization Id",
            "type": "string"
          },
          "organization_name": {
            "title": "Organization Name",
            "type": "string"
          },
          "region": {
            "title": "Region",
            "type": "string"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/StageOut"
            },
            "title": "Stages",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "synthetic": {
            "title": "Synthetic",
            "type": "boolean"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "workflow_id": {
            "title": "Workflow Id",
            "type": "string"
          },
          "workflow_name": {
            "title": "Workflow Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "region",
          "status",
          "synthetic",
          "organization_id",
          "workflow_id",
          "created_at",
          "organization_name",
          "workflow_name",
          "stages"
        ],
        "title": "RunDetail",
        "type": "object"
      },
      "RunSummary": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "organization_id": {
            "title": "Organization Id",
            "type": "string"
          },
          "region": {
            "title": "Region",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "synthetic": {
            "title": "Synthetic",
            "type": "boolean"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "workflow_id": {
            "title": "Workflow Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "region",
          "status",
          "synthetic",
          "organization_id",
          "workflow_id",
          "created_at"
        ],
        "title": "RunSummary",
        "type": "object"
      },
      "StageDefinitionOut": {
        "properties": {
          "action_type": {
            "title": "Action Type",
            "type": "string"
          },
          "approval_roles": {
            "items": {
              "type": "string"
            },
            "title": "Approval Roles",
            "type": "array"
          },
          "escalation_rules": {
            "items": {
              "type": "string"
            },
            "title": "Escalation Rules",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "risk_tier": {
            "title": "Risk Tier",
            "type": "string"
          },
          "sla_hours": {
            "title": "Sla Hours",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "action_type",
          "risk_tier",
          "sla_hours",
          "approval_roles",
          "escalation_rules"
        ],
        "title": "StageDefinitionOut",
        "type": "object"
      },
      "StageOut": {
        "properties": {
          "action_type": {
            "title": "Action Type",
            "type": "string"
          },
          "allowed_roles": {
            "items": {
              "type": "string"
            },
            "title": "Allowed Roles",
            "type": "array"
          },
          "decision": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          },
          "draft": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draft"
          },
          "escalation_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Escalation Reason"
          },
          "evidence": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "risk_tier": {
            "title": "Risk Tier",
            "type": "string"
          },
          "stage_key": {
            "title": "Stage Key",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "stage_key",
          "name",
          "position",
          "action_type",
          "risk_tier",
          "status",
          "allowed_roles"
        ],
        "title": "StageOut",
        "type": "object"
      },
...
      "TemplateDetail": {
        "properties": {
          "actors": {
            "items": {
              "$ref": "#/components/schemas/ActorOut"
            },
            "title": "Actors",
            "type": "array"
          },
          "aliases": {
            "description": "Short names accepted by the API and CLI.",
            "items": {
              "type": "string"
            },
            "title": "Aliases",
            "type": "array"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "key": {
            "description": "Canonical template id, e.g. `retail_refund`.",
            "title": "Key",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "placeholder_schema": {
            "additionalProperties": true,
            "description": "JSON Schema for the typed discovery answers this template needs.",
            "title": "Placeholder Schema",
            "type": "object"
          },
          "sector": {
            "title": "Sector",
            "type": "string"
          },
          "stage_count": {
            "title": "Stage Count",
            "type": "integer"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/StageDefinitionOut"
            },
            "title": "Stages",
            "type": "array"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "key",
          "aliases",
          "name",
          "sector",
          "version",
          "description",
          "stage_count",
          "actors",
          "stages",
          "placeholder_schema"
        ],
        "title": "TemplateDetail",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "contact": {
      "email": "medo433447@gmail.com",
      "name": "Mohamed Mahmoud"
    },
    "description": "\nCall the governed workflow engine from another system.\n\n**What this API will not do.** It will not approve anything on your behalf: a MEDIUM or\nHIGH stage moves only through `POST /runs/{run_id}/stages/{stage_key}/decision`, carrying a\nsigned human decision, and the authenticated principal is written into the hash-chained\nledger beside it. It will also not let a caller choose the model adapter \u2014 that is a\ndeployment setting (invariant **G7**).\n\n**Authentication.** Send `Authorization: Bearer <credential>`, where the credential is\neither a scoped API key (`atmpl_...`, minted with `atmpl keys create`) or a short-lived\naccess token from `POST /oauth/token`. Scopes are enforced per route and listed in each\nroute's errors.\n\n**Errors.** Every failure is `{\"error\": {\"code\", \"message\", \"details\"}}`. Every response\nechoes the `X-Request-Id` you sent, or one generated for you.\n\nAll demo organizations, people, and records reachable through this API are synthetic.\n",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "title": "Enterprise AI Automation Templates API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/runs": {
      "get": {
        "description": "Newest first. Page with the opaque `cursor` returned by the previous page.",
        "operationId": "list_runs_runs_get",
        "parameters": [
          {
            "description": "Filter by run status.",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by run status.",
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "organization_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Organization Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Opaque cursor from `next_cursor`.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor from `next_cursor`.",
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_RunSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List runs",
        "tags": [
          "runs"
        ]
      },
      "post": {
        "description": "Starts a compiled workflow. The model adapter is chosen by this deployment's settings and can never be selected by the caller (G7).",
        "operationId": "create_run_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start a governed run",
        "tags": [
          "runs"
        ]
      }
    },
    "/runs/{run_id}": {
      "get": {
        "operationId": "get_run_runs__run_id__get",
        "parameters": [
          {
            "description": "Run id.",
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "description": "Run id.",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read one run with its stages",
        "tags": [
          "runs"
        ]
      }
    },
    "/runs/{run_id}/stages": {
      "get": {
        "operationId": "list_stages_runs__run_id__stages_get",
        "parameters": [
          {
            "description": "Run id.",
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "description": "Run id.",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StageOut"
                  },
                  "title": "Response List Stages Runs  Run Id  Stages Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List the stages of a run",
        "tags": [
          "runs"
        ]
      }
    },
    "/runs/{run_id}/stages/{stage_key}/decision": {
      "post": {
        "description": "The only path that moves a MEDIUM or HIGH stage. The authenticated principal is written into the hash-chained ledger next to the signature, so the audit answers *who decided*, not only *what was decided*.",
        "operationId": "submit_decision_runs__run_id__stages__stage_key__decision_post",
        "parameters": [
          {
            "description": "Run id.",
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "description": "Run id.",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "description": "Stage key inside that run, e.g. `terminal_decision`.",
            "in": "path",
            "name": "stage_key",
            "required": true,
            "schema": {
              "description": "Stage key inside that run, e.g. `terminal_decision`.",
              "title": "Stage Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record a signed human decision",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/templates/{key}": {
      "get": {
        "description": "Returns the stage graph, the actors, and the JSON Schema of the typed discovery answers the template needs before it can be compiled.",
        "operationId": "get_template_templates__key__get",
        "parameters": [
          {
            "description": "Template id or alias, e.g. `bank` or `loan_triage`.",
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "description": "Template id or alias, e.g. `bank` or `loan_triage`.",
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read one template",
        "tags": [
          "templates"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "This deployment",
      "url": "/api/v1"
    }
  ],
...
}